수신한 패킷을 FrameData로 합치는 PacketCombiner 클래스 작성
- FrameDataManager의 역할 분산을 위해 작성
